feat : add isFavorite for details api

+++ src/main/java/com/iko/iko/service/product/GetAllProductService.java
@@ -57,7 +57,7 @@ public class GetAllProductService {
             }
             ProductDetailsResponse.MainProductForResponse checkData
                     =new ProductDetailsResponse.MainProductForResponse(
-                            isFavorite,
+                    (int)(long)isFavorite,
                     tmp.getProductId(),
                     tmp.getSeries(),
                     gList,
